(refactor) add meta package for enums
