+++ bot/src/main/java/ru/tinkoff/edu/java/bot/bot/CommandsHandler.java
@@ -4,6 +4,7 @@ import com.pengrad.telegrambot.model.Update;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import ru.tinkoff.edu.java.bot.bot.command.CommandHandler;
+import ru.tinkoff.edu.java.bot.bot.meta.State;
 import ru.tinkoff.edu.java.bot.dto.HandledUpdate;
 
 import java.util.List;
