+++ bot/src/main/java/ru/tinkoff/edu/java/bot/bot/MessageHandler.java
@@ -5,6 +5,7 @@ import com.pengrad.telegrambot.request.SendMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import ru.tinkoff.edu.java.bot.bot.command.HelpCommandHandler;
+import ru.tinkoff.edu.java.bot.bot.meta.State;
 import ru.tinkoff.edu.java.bot.dto.HandledUpdate;
 
 import java.util.Optional;
