+++ bot/src/main/java/ru/tinkoff/edu/java/bot/bot/MessageSender.java
@@ -7,6 +7,7 @@ import com.pengrad.telegrambot.model.request.ReplyKeyboardRemove;
 import com.pengrad.telegrambot.request.SendMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import ru.tinkoff.edu.java.bot.bot.meta.Command;
 
 import java.util.Arrays;
