+++ bot/src/main/java/ru/tinkoff/edu/java/bot/bot/UpdatesProcessor.java
@@ -5,6 +5,7 @@ import com.pengrad.telegrambot.UpdatesListener;
 import com.pengrad.telegrambot.model.Update;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import ru.tinkoff.edu.java.bot.bot.meta.State;
 import ru.tinkoff.edu.java.bot.dto.HandledUpdate;
 
 import java.util.HashMap;
