+++ bot/src/main/java/ru/tinkoff/edu/java/bot/bot/command/CommandHandler.java
@@ -3,8 +3,8 @@ package ru.tinkoff.edu.java.bot.bot.command;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.BaseRequest;
 import com.pengrad.telegrambot.response.BaseResponse;
-import ru.tinkoff.edu.java.bot.bot.Command;
-import ru.tinkoff.edu.java.bot.bot.State;
+import ru.tinkoff.edu.java.bot.bot.meta.Command;
+import ru.tinkoff.edu.java.bot.bot.meta.State;
 
 public interface CommandHandler<T extends BaseRequest<T, R>, R extends BaseResponse> {
     Command command();
