+++ bot/src/main/java/ru/tinkoff/edu/java/bot/bot/command/HelpCommandHandler.java
@@ -2,14 +2,13 @@ package ru.tinkoff.edu.java.bot.bot.command;
 
 
 import com.pengrad.telegrambot.model.Update;
-import com.pengrad.telegrambot.request.BaseRequest;
 import com.pengrad.telegrambot.request.SendMessage;
 import com.pengrad.telegrambot.response.SendResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import ru.tinkoff.edu.java.bot.bot.Command;
+import ru.tinkoff.edu.java.bot.bot.meta.Command;
 import ru.tinkoff.edu.java.bot.bot.MessageSender;
-import ru.tinkoff.edu.java.bot.bot.State;
+import ru.tinkoff.edu.java.bot.bot.meta.State;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
