+++ bot/src/main/java/ru/tinkoff/edu/java/bot/dto/HandledUpdate.java
@@ -2,7 +2,7 @@ package ru.tinkoff.edu.java.bot.dto;
 
 import com.pengrad.telegrambot.request.BaseRequest;
 import lombok.Builder;
-import ru.tinkoff.edu.java.bot.bot.State;
+import ru.tinkoff.edu.java.bot.bot.meta.State;
 
 import java.util.Optional;
